Update request user-agent
The previous one was really old and I wouldn't be surprised if BeReal
started blocking account using that after the c&d email sent to
the original author of the project.

The auth requests are still sent using iOS keys, however I was only able
to get a more recent user-agent from an Android phone. The auth tokens
are interchangeable across platforms (at least for now).